Added condition to copy file.

+++ build/libraries/build.glib.sh
@@ -5,4 +5,6 @@ cd build
 export CFLAGS="$FLAGS"
 meson --prefix=/usr/local --default-library=static -Dselinux=disabled -Dxattr=false -Dlibmount=disabled -Dnls=disabled -Dinternal_pcre=true -Dgtk_doc=false -Dman=false ..
 ninja install
-cp /usr/local/lib/x86_64-linux-gnu/libglib-2.0.a /usr/local/lib/libglib-2.0.a
+if [ -f /usr/local/lib/x86_64-linux-gnu/libglib-2.0.a ]; then
+  cp /usr/local/lib/x86_64-linux-gnu/libglib-2.0.a /usr/local/lib/libglib-2.0.a
+fi
